% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Graphs.R
\name{compare_graphs}
\alias{compare_graphs}
\title{Plots networks from several models}
\usage{
compare_graphs(allNets, curv = 0.2, width = 1, shade = TRUE,
  Ft = 0, nodes_label = NULL, btw_rank = 2, layout = "circle",
  base_model = NULL)
}
\arguments{
\item{allNets}{tibble resulting of ComparEMtree()}

\item{curv}{edges curvature}

\item{width}{maximum width for the edges}

\item{shade}{if TRUE, shades the edges non-linked to nodes with high betweenness}

\item{Ft}{Frequency threshold}

\item{nodes_label}{optional label for the nodes}

\item{btw_rank}{betweenness rank +1 of highlighted nodes. If set to 1, none are highlighted.}

\item{layout}{choice of layout for all networks among available layouts in `ggraph`. Default is "circle"}

\item{base_model}{choice of referent model for the layout construction}
}
\value{
\itemize{
\item{G}{ the collection of networks as a ggplot2 object, with highlighted high betweenness nodes}
\item{graph_data }{list of data needed to plot the networks}
}
}
\description{
Plots networks from several models
}
\examples{
n=30
p=10
S=3
Y=data_from_scratch("tree",p=p,n=n)$data
X = data.frame(rnorm(n),rbinom(n,1,0.7))
threemodels=ComparEMtree(Y,X,models=list(1,2,c(1,2)),
m_names=list("1","2","both"),Pt=0.3,S=S, cores=1)
compare_graphs(threemodels)
}
