% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener_data.R
\name{data_from_scratch}
\alias{data_from_scratch}
\title{Simulates data under the PLN model with control on the dependency structure}
\usage{
data_from_scratch(
  type,
  p = 20,
  n = 50,
  r = 5,
  covariates = NULL,
  dens = log(p)/p,
  k = 3,
  norm = FALSE,
  signed = FALSE,
  v = 0.01,
  draw = FALSE
)
}
\arguments{
\item{type}{type of graph, either "tree", "erdos", "cluster" or "scale-free"}

\item{p}{wanted number of columns (species)}

\item{n}{wanted number of rows (samples/observations)}

\item{r}{within/between connectivity ratio for cluster graphs}

\item{covariates}{a data.frame or matrix containing data covariates.}

\item{dens}{density of edges for cluster graphs or edge probability for Erdös graphs}

\item{k}{number of groups for the cluster structure.}

\item{norm}{should the Gaussian parameters be normalized ?}

\item{signed}{boolean: should the graph be composed of positive and negative partial correlations ?}

\item{v}{noise parameter of the precision matrix}

\item{draw}{boolean, plots the graph if set to TRUE}
}
\value{
a list containing
\itemize{
 \item{data: }{simulated counts}
 \item{omega: }{the precision matrix}
}
}
\description{
Simulates data under the PLN model with control on the dependency structure
}
\examples{
set.seed(1)
p<-30
Y1<-data_from_scratch("tree",p=p,draw=TRUE)
str(Y1)
Y2<-data_from_scratch("cluster",p=p,r=10, dens=10/p, k=3,draw=TRUE)
}
