% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener_data.R
\name{erdos}
\alias{erdos}
\title{Simulates an Erdös-Renyi graph}
\usage{
erdos(p, prob)
}
\arguments{
\item{p}{number of nodes}

\item{prob}{probability for an edge, which is the same for all edges}
}
\value{
the adjacency matrix
}
\description{
Simulates an Erdös-Renyi graph
}
\examples{
erdos(10,0.3)
}
