% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{FitBeta}
\alias{FitBeta}
\title{EMtree core computing function}
\usage{
FitBeta(beta.init, psi, maxIter = 50, eps1 = 1e-06, eps2 = 1e-04,
  verbatim = TRUE, plot = FALSE)
}
\arguments{
\item{beta.init}{initialization of beta weights}

\item{psi}{psi matrix, filled with ratios of bivariate probabilities over marginals, which can in the Gaussian
wase be deduced from the correlation matrix.}

\item{maxIter}{maximum number of iterations}

\item{eps1}{higher bound of beta weights variation}

\item{eps2}{higher bound of log likelihood variation}

\item{verbatim}{boolean for verbosity}

\item{plot}{boolean for plotting of the likelihood trajectory}
}
\value{
\itemize{
 \item{edges_prob: }{p x p matrix of edges probabilities}
 \item{edges_weight: }{p x p matrix of edges weights for any spanning tree}
 \item{logpY: }{vector of log-likelihoods}
 \item{maxIter: }{final number of iterations EMtree has ran}
 \item{timeEM: }{EMtree computation time}
}
}
\description{
EMtree core computing function
}
\examples{
n=30
p=10
S=5
Y=data_from_scratch("tree",p=p,n=n)$data
beta = matrix(1/10,10,10)
psi=Psi_alpha(cor(Y), n)$psi
FitEM = FitBeta(beta.init=beta, psi=psi, maxIter = 6, verbatim=TRUE, plot=TRUE)
}
