% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{FitBetaStatic}
\alias{FitBetaStatic}
\title{internal function}
\usage{
FitBetaStatic(beta.init, psi, maxIter = 50, eps1 = 1e-06,
  eps2 = 1e-04, verbatim = TRUE, plot = FALSE)
}
\arguments{
\item{plot}{}
}
\description{
internal function
}
