% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{freq_selec}
\alias{freq_selec}
\title{Computes edges selection frequency after resampling procedure}
\usage{
freq_selec(Pmat, Pt)
}
\arguments{
\item{Pmat}{matrix gathering edgges probability, with dimensions number of resamples x number of possible edges. Typically the Pmat output of ResampleEMtree()}

\item{Pt}{edges probabilty threshold}
}
\value{
matrix of edges selection frequency
}
\description{
Computes edges selection frequency after resampling procedure
}
\examples{
n=30
p=10
S=5
Y=data_from_scratch("tree",p=p)$data
resample_prob=ResampleEMtree(Y, S=S,cores = 1)$Pmat
edges_freq<-freq_selec(resample_prob,Pt=0.3)
str(edges_freq)
}
