% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener_data.R
\name{generator_PLN}
\alias{generator_PLN}
\title{Simulate count data under the Poisson log-Normal model}
\usage{
generator_PLN(Sigma, covariates = NULL, n = 50, norm = FALSE)
}
\arguments{
\item{Sigma}{Covariance matrix of the normal hidden layer of parameters}

\item{covariates}{a data.frame or matrix containing data covariates. If not NULL, defines the
number of  simulated rows.}

\item{n}{number of rows to simulate}

\item{norm}{should the parameters be normalized ?}
}
\value{
\itemize{
\item{if norm=FALSE}{ Y: the simulated counts}
\item{if norm=TRUE}{ \itemize{
      \item{Y:}{simulated counts}
      \item{U:}{ the normalized Gaussian parameters}}}}
}
\description{
Simulate count data under the Poisson log-Normal model
}
\examples{
G=generator_graph(p=10,graph="tree")
sigma=generator_param(G=G)$sigma
generator_PLN(as.matrix(sigma))
}
