% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsMatVec.R
\name{inverse.gmp}
\alias{inverse.gmp}
\title{Computes exact inverses using gmp package}
\usage{
inverse.gmp(A)
}
\arguments{
\item{A}{squared positive definite matrix}
}
\value{
the exact inverse of A
}
\description{
Computes exact inverses using gmp package
}
