% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{Psi_alpha}
\alias{Psi_alpha}
\title{Choice of alpha for a numerically stable psi matrix}
\usage{
Psi_alpha(CorY, n, cond.tol = 1e-10)
}
\arguments{
\item{CorY}{Correlation matrix of dimensoins p x p with p the number of species for example.}

\item{n}{number of samples}

\item{cond.tol}{tolerance for the matrix conditionment, which is here the ratio of the absolute values of minimal eigen value on maximal eigen value}
}
\value{
\itemize{
\item{psi: }{p x p corrected psi matrix}
\item{alpha: }{final alpha value needed to correct the psi matrix}}
}
\description{
Choice of alpha for a numerically stable psi matrix
}
\examples{
n=30
p=10
S=5
Y=data_from_scratch("tree",p=p,n=n)$data
beta = matrix(1/10,10,10)
psi=Psi_alpha(cor(Y), n)$psi
}
