% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener_data.R
\name{SimCluster}
\alias{SimCluster}
\title{Simulates a cluster graph}
\usage{
SimCluster(p, k, dens, r)
}
\arguments{
\item{p}{number of nodes}

\item{k}{number of groups in the cluster}

\item{dens}{graph density}

\item{r}{within/between ratio connection probability}
}
\value{
the adjacency matrix of size p x p
}
\description{
Simulates a cluster graph
}
\examples{
SimCluster(10,2,0.5, 10)
}
