% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener_data.R
\name{SpannTree}
\alias{SpannTree}
\title{Simulates a spanning tree graph from `vegan`package}
\usage{
SpannTree(p)
}
\arguments{
\item{p}{number of nodes}
}
\value{
the adjacency matrix
}
\description{
Simulates a spanning tree graph from `vegan`package
}
\examples{
SpannTree(10)
}
