% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{StATS}
\alias{StATS}
\title{Stability Approach to Threshold Selection}
\usage{
StATS(Pmat, nlambda = 100, stab.thresh = 0.9, plot = FALSE)
}
\arguments{
\item{Pmat}{Pmat output from the `ResampleEMtree()` function}

\item{nlambda}{Number of probability thresholds to look at}

\item{stab.thresh}{Stability threshold}

\item{plot}{Optional graphic output}
}
\value{
a list containing
\itemize{
 \item{freqs_opt: }{A tibble with (number of possible edges) x nlambda rows, containing the computed selection frequency of each edge and stability measure for each threshold.}
 \item{lambda_opt: }{The probability threshold giving the desired stability of frequencies.}
 \item{plot: }{The optional graphic output.}
}
}
\description{
Stability Approach to Threshold Selection
}
\examples{
n=100
p=15
S=15
set.seed(2021)
simu=data_from_scratch("erdos",p=p,n=n)
G=1*(simu$omega!=0) ; diag(G) = 0
#With default evaluation, using the PLNmodel paradigm:
default_resample=ResampleEMtree(simu$data, S=S, cores = 1)
stab_selection=StATS(default_resample$Pmat, nlambda=50, stab.thresh=0.9,plot=TRUE)
#Check quality of result
table(pred=1*(stab_selection$freqs_opt>0.9), truth=ToVec(G))
}
