% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsTree.R
\name{SumTree}
\alias{SumTree}
\title{Summing over trees (Matrix-tree thm)}
\usage{
SumTree(W)
}
\arguments{
\item{W}{Squared weight matrix}
}
\value{
The number of total possible spanning trees
}
\description{
Summing over trees (Matrix-tree thm)
}
\examples{
uniformeW=matrix(1,3,3)
SumTree(uniformeW)
}
