% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsMatVec.R
\name{ToSym}
\alias{ToSym}
\title{Makes a symmetric matrix from the vector made of its upper triangular part}
\usage{
ToSym(A.vec)
}
\arguments{
\item{A.vec}{Vector obtain with ToVec or the upper.tri() function.}
}
\value{
The symmetric matrix with A.vec as off-diagonal terms.
}
\description{
Makes a symmetric matrix from the vector made of its upper triangular part
}
