% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsMatVec.R
\name{ToVec}
\alias{ToVec}
\title{Makes a vector from the upper triangular part of a symmetric matrix}
\usage{
ToVec(A.mat)
}
\arguments{
\item{A.mat}{A symmetric matrix.}
}
\value{
The vector from the upper triangular part of A.mat.
}
\description{
Makes a vector from the upper triangular part of a symmetric matrix
}
