% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{ComparEMtree}
\alias{ComparEMtree}
\title{Title}
\usage{
ComparEMtree(counts, vec_covar, O = NULL, v = 0.8, B = 100, maxIter,
  cond.tol = 1e-14, cores = 3, f)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
