% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener_data.R
\name{data_from_scratch}
\alias{data_from_scratch}
\title{generates data under the PLN model with a certain type of dependency structure, and draws the structure.}
\usage{
data_from_scratch(type, p = 20, n = 50, r = 5, covariates = NULL,
  prob = log(p)/p, dens = log(p)/p, draw = FALSE)
}
\arguments{
\item{type}{type of graph, either "tree", "erdos", "cluster" or "scale-free"}

\item{p}{wanted number of columns (species)}

\item{r}{within/between connectiviy ratio for cluster graphs}

\item{covariates}{a data.frame or matrix containing data covariates.}

\item{prob}{edge probability for erdos graphs}

\item{dens}{density of edges for cluster graphs}

\item{draw}{boolean, plots the graph if set to TRUE}
}
\value{
a list containing
\itemize{
 \item{data: }{simulated counts}
 \item{omega: }{the precision matrix}
}
}
\description{
generates data under the PLN model with a certain type of dependency structure, and draws the structure.
}
\examples{
data_from_scratch("tree",p=10,draw=TRUE)
}
