% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Graphs.R
\name{draw_network}
\alias{draw_network}
\title{Title}
\usage{
draw_network(adj_matrix, title = "", size = 4, curv = 0.2,
  filter_deg = FALSE, layout = NULL, nodes_label = NULL,
  pal = NULL, seed = 200)
}
\arguments{
\item{adj_matrix}{graph adjacency matrix}

\item{title}{graph title}

\item{size}{size of nodes}

\item{curv}{edges curvature}

\item{filter_deg}{selects nodes with a higher degree than filter_deg}

\item{layout}{optional igraph layout.}

\item{nodes_label}{optional labels for nodes.}

\item{pal}{optional palette.}

\item{seed}{optional seed for graph reproductibility.}
}
\value{
plots a network
}
\description{
Title
}
\examples{
adj_matrix= SimCluster(10,2,0.5, 0.1)
draw_network(adj_matrix,"Cluster graph", layout="kk")
}
