% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_Graphs.R
\name{draw_network}
\alias{draw_network}
\title{Plots a network}
\usage{
draw_network(
  adj_matrix,
  title = "",
  size = 4,
  curv = 0,
  width = 1,
  shade = FALSE,
  filter_deg = FALSE,
  btw_rank = 2,
  layout = NULL,
  stored_layout = NULL,
  nodes_label = NULL,
  nodes_size = c(2, 5),
  pal_edges = NULL,
  pal_nodes = NULL,
  groupes = NULL
)
}
\arguments{
\item{adj_matrix}{graph adjacency matrix}

\item{title}{graph title}

\item{size}{size of nodes}

\item{curv}{edges curvature}

\item{width}{maximum width for the edges}

\item{shade}{if TRUE, shades the edges unlinked to nodes with high betweenness}

\item{filter_deg}{selects nodes with a higher degree than filter_deg}

\item{btw_rank}{betweenness rank +1 of highlighted nodes. If set to 1, none are highlighted.}

\item{layout}{optional ggraph layout.}

\item{stored_layout}{optional data.frame of point positions from a previous graph, possibly created using ggraph::create_layout()}

\item{nodes_label}{optional labels for nodes.}

\item{nodes_size}{size of nodes, possibility to specify a size per group}

\item{pal_edges}{optional palette for edges}

\item{pal_nodes}{optional palette for nodes}

\item{groupes}{optional vector seperating the nodes into groupes}
}
\value{
\itemize{
\item{G} {the network as a ggplot2 object, with highlighted high betweenness nodes}
\item{graph_data}{ data needed for plotting the network}
}
}
\description{
Plots a network
}
\examples{
adj_matrix= SimCluster(20,2,0.4, 10)
draw_network(adj_matrix,"Cluster graph", layout="fr", shade=TRUE)
}
