% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsTree.R
\name{EdgeProba}
\alias{EdgeProba}
\title{Computing edges conditional probabilities}
\usage{
EdgeProba(W, verbatim = FALSE)
}
\arguments{
\item{W}{squared weight matrix}

\item{verebatim}{controls verbosity}
}
\value{
Edges conditional probabilities computed directly, without using the Kirshner function.
}
\description{
Computing edges conditional probabilities
}
\examples{
W = matrix(c(1,1,3,1,1,1,3,1,1),3,3,byrow=TRUE)
EdgeProba(W)
}
