% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{EMtree}
\alias{EMtree}
\title{Computes edges probability}
\usage{
EMtree(PLN.Cor, maxIter = 30, cond.tol = 1e-10, PLN = TRUE,
  verbatim = TRUE, plot = FALSE)
}
\arguments{
\item{PLN.Cor}{Either an object resulting from the use of the `PLN` function from package `PLNmodels`, or an estimation of Gaussian data correlation matrix.}

\item{maxIter}{Maximum number of iterations for EMtree}

\item{cond.tol}{Tolerence parameter for the conditionement of psi matrix}

\item{verbatim}{Talks if set to TRUE}

\item{plot}{Plots likelihood if set to TRUE}
}
\value{
\itemize{
 \item{edges_prob: }{p x p matrix of edges probabilities}
 \item{edges_weight: }{p x p matrix of edges weights for any spanning tree}
 \item{logpY: }{vector of log-likelihoods}
 \item{maxIter: }{final number of iterations EMtree has ran}
 \item{timeEM: }{EMtree computation time}
 \item{alpha: }{data signal/noise ratio}
}
}
\description{
Computes edges probability
}
\examples{
n=30
p=10
Y=data_from_scratch("tree",p=p)$data
PLN_Y = PLNmodels::PLN(Y~1)
CorY=cov2cor(PLN_Y$model_par$Sigma)
EMtree(PLN.Cor=PLN_Y,verbatim=TRUE, plot=TRUE)
}
