% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{FitBeta}
\alias{FitBeta}
\title{Update the beta weights according to the gradient ascent}
\usage{
FitBeta(beta.init, psi, maxIter = 50, eps = 1e-06, unlinked = NULL)
}
\arguments{
\item{beta.init}{Initial beta weight matrix}

\item{psi}{Psi matrix, filled with ratios of bivariate probabilities over marginals, which can in the Gaussian
wase be deduced from the correlation matrix.}

\item{maxIter}{Maximum number of iterations}

\item{eps}{Precision parameter controlling the convergence of weights beta}

\item{unlinked}{An optional vector of nodes which are not linked with each other}
}
\value{
\itemize{
 \item{edges_prob: }{p x p matrix of edges probabilities}
 \item{edges_weight: }{p x p matrix of edges weights for any spanning tree}
 \item{logpY: }{vector of log-likelihoods}
 \item{maxIter: }{final number of iterations EMtree has ran}
 \item{timeEM: }{EMtree computation time}
}
}
\description{
Update the beta weights according to the gradient ascent
}
\examples{
n=30
p=10
S=5
Y=data_from_scratch("tree",p=p,n=n)$data
beta = matrix(1/10,10,10)
psi=Psi_alpha(cor(Y), n)$psi
FitEM = FitBeta(beta.init=beta, psi=psi, maxIter = 6)
}
