% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{FitBetaStatic}
\alias{FitBetaStatic}
\title{internal function}
\usage{
FitBetaStatic(beta.init, psi, maxIter = 50, eps1 = 1e-06,
  eps2 = 1e-04, verbatim = TRUE, plot = FALSE)
}
\arguments{
\item{beta.init}{initialization of beta weights}

\item{psi}{psi matrix, filled with ratios of bivariate probabilities over marginals, which can in the Gaussian
wase be deduced from the correlation matrix.}

\item{maxIter}{maximum number of iterations}

\item{eps1}{variation higher bound of beta weights}

\item{eps2}{variation higher bound of log likelihood}

\item{verbatim}{boolean}

\item{plot}{boolean}
}
\description{
internal function
}
