% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gener_data.R
\name{generator_param}
\alias{generator_param}
\title{Generate simulation parameters from a graph adjacency matrix}
\usage{
generator_param(G, signed = FALSE, v = 0.01)
}
\arguments{
\item{G}{adjacency matrix}

\item{signed}{boolean: should the graph be composed of positive and negative partial correlations ?}

\item{v}{parameter controlling the noise on the precision matrix}
}
\value{
a list containing
\itemize{
 \item{sigma: }{the covariance matrix}
 \item{omega: }{the precision matrix}
 \item{cste: }{the constant that was needed to load the diagonal of omega and esure its positive definitiveness}
}
}
\description{
Generate simulation parameters from a graph adjacency matrix
}
\examples{
G=generator_graph(p=10,graph="tree")
generator_param(G=G)

}
