% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsTree.R
\name{Kirshner}
\alias{Kirshner}
\title{Computing edges marginal probabilities}
\usage{
Kirshner(W)
}
\arguments{
\item{W}{Squared weight matrix}
}
\value{
Edges probabilities as defined in Kirshner 2007
}
\description{
Computing edges marginal probabilities
}
\examples{
W = matrix(c(1,1,3,1,1,1,3,1,1),3,3,byrow=TRUE)
Kirshner(W)
}
