% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{ResampleEMtree}
\alias{ResampleEMtree}
\title{Title}
\usage{
ResampleEMtree(counts, vec_covar, O = NULL, v = 0.8, B = 100,
  maxIter, cond.tol = 1e-14, cores = 3)
}
\arguments{
\item{cores}{}
}
\description{
Title
}
