% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{ResampleEMtree}
\alias{ResampleEMtree}
\title{Resampling procedure for  edges probability}
\usage{
ResampleEMtree(
  counts,
  covar_matrix = NULL,
  unlinked = NULL,
  O = NULL,
  user_covariance_estimation = NULL,
  v = 0.8,
  S = 100,
  maxIter = 30,
  cond.tol = 1e-10,
  eps = 0.001,
  cores = 3,
  init = FALSE
)
}
\arguments{
\item{counts}{Data of observed counts with dimensions n x p, either a matrix, data.frame or tibble.}

\item{covar_matrix}{matrix of covariates, should have the same number of rows as the count matrix.}

\item{unlinked}{An optional vector of nodes which are not linked with each other}

\item{O}{Matrix of offsets, with dimension n x p}

\item{user_covariance_estimation}{A user-provided function for the estimation of a covariance}

\item{v}{The proportion of observed data to be taken in each sub-sample. It is the ratio (sub-sample size)/n}

\item{S}{Total number of wanted sub-samples.}

\item{maxIter}{Maximum number of EMtree iterations at each sub-sampling.}

\item{cond.tol}{Tolerance for the psi matrix.}

\item{eps}{Precision parameter controlling the convergence of weights beta}

\item{cores}{Number of cores, can be greater than 1 if data involves less than about 32 species.}

\item{init}{boolean: should the resampling be carried out with different initial points (TRUE), or with different initial data (FALSE)}
}
\value{
Returns a list which contains the Pmat data.frame, and vectors of EMtree maximum iterations and running times in each
resampling.
\itemize{
 \item{Pmat: }{S x p(p-1)/2 matrix with edge probabilities for each resample}
 \item{maxIter: }{EMtree maximum iterations in each resampling.}
 \item{times: }{EMtree running times in each resampling.}
}
}
\description{
Resampling procedure for  edges probability
}
\examples{
n=100
p=12
S=5
set.seed(2021)
simu=data_from_scratch("erdos",p=p,n=n)
G=1*(simu$omega!=0) ; diag(G) = 0
# With default evaluation, using the PLNmodel paradigm:
default_resample=ResampleEMtree(simu$data, S=S,cores = 1)

# With provided correlation estimation function:
estimSigma<-function(counts, covar_matrix, sample){
Dum_Sigma = cov2cor(cov(counts[sample,]))
}
custom_resample=ResampleEMtree(simu$data,S=S,cores = 1,user_covariance_estimation=estimSigma)

# We then run the stability selection to find the optimal selection frequencies,
# for a stability of 85\%:
stab_default=StATS(default_resample$Pmat, nlambda=50, stab.thresh=0.8,plot=TRUE)
stab_custom=StATS(custom_resample$Pmat, nlambda=50, stab.thresh=0.8,plot=TRUE)

#Check quality of result
table(pred=1*(stab_default$freqs_opt>0.9), truth=ToVec(G))
table(pred=1*(stab_custom$freqs_opt>0.9), truth=ToVec(G))
}
