% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_inference.R
\name{ResampleEMtree}
\alias{ResampleEMtree}
\title{Resampling procedure for  edges probability}
\usage{
ResampleEMtree(
  counts,
  covar_matrix = NULL,
  O = NULL,
  v = 0.8,
  S = 100,
  maxIter = 30,
  cond.tol = 1e-14,
  cores = 3
)
}
\arguments{
\item{counts}{Data of observed counts with dimensions n x p, either a matrix, data.frame or tibble.}

\item{covar_matrix}{matrix of covariates, should have the same number of rows as the count matrix.}

\item{O}{Matrix of offsets, with dimension n x p}

\item{v}{The proportion of observed data to be taken in each sub-sample. It is the ratio (sub-sample size)/n}

\item{S}{Total number of wanted sub-samples.}

\item{maxIter}{Maximum number of EMtree iterations at each sub-sampling.}

\item{cond.tol}{Tolerance for the psi matrix.}

\item{cores}{Number of cores, can be greater than 1 if data involves less than about 32 species.}
}
\value{
Returns a list which contains the Pmat data.frame, and vectors of EMtree maximum iterations and running times in each
resampling.
\itemize{
 \item{Pmat: }{S x p(p-1)/2 matrix with edge probabilities for each resample}
 \item{maxIter: }{EMtree maximum iterations in each resampling.}
 \item{times: }{EMtree running times in each resampling.}
}
}
\description{
Resampling procedure for  edges probability
}
\examples{
n=30
p=10
S=5
Y=data_from_scratch("tree",p=p,n=n)$data
X = data.frame(rnorm(n),rbinom(n,1,0.7))
resample=ResampleEMtree(Y,covar_matrix=X, S=S,cores = 1)
str(resample)
}
